Committing final updates to admin form and subforms, plus last minute StyleCop fixes.
Resized AdminForm window so it works well with both subforms.  Also updated the Change Password buttons on UserAdminForm so they open the correct forms.

Added validation to the UpdatePasswordForm (admin version) and changed the constructor to receive a UserID string as input.  This allows admins to change the password without knowing the current one.